Squashed commit of the following:

commit 54354ed36b8aca6b60c96545963fe64206b91f43
Author: DEXREQ Automation <noreply+dexreq@oracle.com>
Date:   Fri May 29 18:20:16 2020 +0000

    Author: DEXREQ Automation <noreply+dexreq@oracle.com>
    Date:  Fri, 29 May 2020 18:20:16 +0000
    Running generation for: [[DEXREQ-1315, DEXREQ-1310, DEXREQ-1313, DEXREQ-1301, DEXREQ-1312]]

    Author: DEXREQ Automation <noreply+dexreq@oracle.com>
    Date:  Fri, 29 May 2020 17:45:32 +0000
    Updating pom.xml for [[DEXREQ-1315, DEXREQ-1310, DEXREQ-1313, DEXREQ-1301, DEXREQ-1312]]

    (automatically combined)

+++ database/exadata_infrastructure.go
@@ -56,7 +56,7 @@ type ExadataInfrastructure struct {
 	DataStorageSizeInTBs *float64 `mandatory:"false" json:"dataStorageSizeInTBs"`
 
 	// The total available DATA disk group size.
-	MaxDataStorageInTBs *int `mandatory:"false" json:"maxDataStorageInTBs"`
+	MaxDataStorageInTBs *float64 `mandatory:"false" json:"maxDataStorageInTBs"`
 
 	// The IP address for the first control plane server.
 	CloudControlPlaneServer1 *string `mandatory:"false" json:"cloudControlPlaneServer1"`
